Phase 3 starts. Kafka initial implementation. Notification Service.

+++ backend/orders/src/main/java/com/olomsky/orders/dto/OrderResponse.java
@@ -1,10 +1,8 @@
 package com.olomsky.orders.dto;
 
-import java.math.BigDecimal;
-
 public record OrderResponse(
     Long id,
     String orderNumber,
     String skuCode,
-    BigDecimal price,
+    Double price,
     Integer quantity) {}
